Refactor MTE-183 [v109] M1 specific test failures in DownloadFilesTests, HomePageSettingsUITests and ReaderViewTest (#12449)
* Avoid swipeLeft() on M1. Use TIMEOUT and TIMEOUT_LONG for timeout values.

* Eliminate print statement :p

* Update testCheckHomeSettingsByDefault to the current state of the settings

* Mitigate intermittent failure on textField not showing input

* Oops. There's no timeout arg for waitForValueContains. Let's add waitForNoExistence to ensure there's enough time.

* Disable testClipboard() because clipboard may not work on M1.

+++ Tests/XCUITests/DownloadFilesTests.swift
@@ -12,30 +12,40 @@ class DownloadFilesTests: BaseTestCase {
     override func tearDown() {
         // The downloaded file has to be removed between tests
         waitForExistence(app.tables["DownloadsTable"])
-
-        let list = app.tables["DownloadsTable"].cells.count
-        if list != 0 {
-            for _ in 0...list-1 {
-                waitForExistence(app.tables["DownloadsTable"].cells.element(boundBy: 0))
-                app.tables["DownloadsTable"].cells.element(boundBy: 0).swipeLeft()
-                waitForExistence(app.tables.cells.buttons["Delete"])
-                app.tables.cells.buttons["Delete"].tap()
+        if processIsTranslatedStr() == m1Rosetta {
+            navigator.nowAt(LibraryPanel_Downloads)
+            navigator.goto(NewTabScreen)
+            navigator.performAction(Action.CloseURLBarOpen)
+            navigator.nowAt(NewTabScreen)
+            navigator.goto(ClearPrivateDataSettings)
+            app.cells.switches["Downloaded Files"].tap()
+            app.tables.cells["ClearPrivateData"].tap()
+            app.alerts.buttons["OK"].tap()
+        } else {
+            let list = app.tables["DownloadsTable"].cells.count
+            if list != 0 {
+                for _ in 0...list-1 {
+                    waitForExistence(app.tables["DownloadsTable"].cells.element(boundBy: 0))
+                    app.tables["DownloadsTable"].cells.element(boundBy: 0).swipeLeft()
+                    waitForExistence(app.tables.cells.buttons["Delete"])
+                    app.tables.cells.buttons["Delete"].tap()
+                }
             }
         }
     }
 
     private func deleteItem(itemName: String) {
         app.tables.cells.staticTexts[itemName].swipeLeft()
-        waitForExistence(app.tables.cells.buttons["Delete"], timeout: 3)
+        waitForExistence(app.tables.cells.buttons["Delete"], timeout: TIMEOUT)
         app.tables.cells.buttons["Delete"].tap()
     }
 
     func testDownloadFilesAppMenuFirstTime() {
-        waitForExistence(app.buttons["urlBar-cancel"], timeout: 5)
+        waitForExistence(app.buttons["urlBar-cancel"], timeout: TIMEOUT)
         navigator.performAction(Action.CloseURLBarOpen)
         navigator.nowAt(NewTabScreen)
         navigator.goto(LibraryPanel_Downloads)
-        waitForExistence(app.tables["DownloadsTable"], timeout: 5)
+        waitForExistence(app.tables["DownloadsTable"], timeout: TIMEOUT)
         XCTAssertTrue(app.tables["DownloadsTable"].exists)
         // Check that there is not any items and the default text shown is correct
         checkTheNumberOfDownloadedItems(items: 0)
@@ -46,9 +56,13 @@ class DownloadFilesTests: BaseTestCase {
         navigator.openURL(testURL)
         waitUntilPageLoad()
         // Verify that the context menu prior to download a file is correct
+        if !iPad() {
+            app.webViews.links.firstMatch.swipeLeft(velocity: 1000)
+            app.webViews.links.firstMatch.swipeLeft(velocity: 1000)
+        }
         app.webViews.links[testFileName].firstMatch.tap()
 
-        waitForExistence(app.tables["Context Menu"], timeout: 5)
+        waitForExistence(app.tables["Context Menu"], timeout: TIMEOUT)
         XCTAssertTrue(app.tables["Context Menu"].staticTexts[testFileNameDownloadPanel].exists)
         XCTAssertTrue(app.tables["Context Menu"].otherElements["download"].exists)
         app.buttons["Cancel"].tap()
@@ -63,7 +77,7 @@ class DownloadFilesTests: BaseTestCase {
         navigator.goto(BrowserTabMenu)
         navigator.goto(LibraryPanel_Downloads)
 
-        waitForExistence(app.tables["DownloadsTable"], timeout: 5)
+        waitForExistence(app.tables["DownloadsTable"], timeout: TIMEOUT)
         // There should be one item downloaded. It's name and size should be shown
         checkTheNumberOfDownloadedItems(items: 1)
         XCTAssertTrue(app.tables.cells.staticTexts[testFileNameDownloadPanel].exists)
@@ -83,26 +97,32 @@ class DownloadFilesTests: BaseTestCase {
         XCTAssertTrue(app.tables.cells.staticTexts[testBLOBFileSize].exists)
     }
 
-    func testDeleteDownloadedFile() {
+    func testDeleteDownloadedFile() throws {
         downloadFile(fileName: testFileName, numberOfDownlowds: 1)
         navigator.goto(BrowserTabMenu)
         navigator.goto(LibraryPanel_Downloads)
         waitForExistence(app.tables["DownloadsTable"])
-
-        deleteItem(itemName: testFileNameDownloadPanel)
-        waitForNoExistence(app.tables.cells.staticTexts[testFileNameDownloadPanel])
-
-        // After removing the number of items should be 0
-        checkTheNumberOfDownloadedItems(items: 0)
+        if processIsTranslatedStr() == m1Rosetta {
+            throw XCTSkip("swipeLeft() does not work on M1")
+        } else {
+            deleteItem(itemName: testFileNameDownloadPanel)
+            waitForNoExistence(app.tables.cells.staticTexts[testFileNameDownloadPanel])
+            // After removing the number of items should be 0
+            checkTheNumberOfDownloadedItems(items: 0)
+        }
     }
 
-    func testShareDownloadedFile() {
+    func testShareDownloadedFile() throws {
         downloadFile(fileName: testFileName, numberOfDownlowds: 1)
         navigator.goto(BrowserTabMenu)
         navigator.goto(LibraryPanel_Downloads)
-        app.tables.cells.staticTexts[testFileNameDownloadPanel].swipeLeft()
-        XCTAssertTrue(app.tables.buttons.staticTexts["Share"].exists)
-        XCTAssertTrue(app.tables.buttons.staticTexts["Delete"].exists)
+        if processIsTranslatedStr() == m1Rosetta {
+            throw XCTSkip("swipeLeft() does not work on M1")
+        } else {
+            app.tables.cells.staticTexts[testFileNameDownloadPanel].swipeLeft()
+            XCTAssertTrue(app.tables.buttons.staticTexts["Share"].exists)
+            XCTAssertTrue(app.tables.buttons.staticTexts["Delete"].exists)
+        }
     }
 
     func testLongPressOnDownloadedFile() {
@@ -113,9 +133,15 @@ class DownloadFilesTests: BaseTestCase {
         waitForExistence(app.tables["DownloadsTable"])
         // Commenting out until share sheet can be managed with automated tests issue #5477
         app.tables.cells.staticTexts[testFileNameDownloadPanel].press(forDuration: 2)
-        waitForExistence(app.otherElements["ActivityListView"], timeout: 10)
+        waitForExistence(app.otherElements["ActivityListView"], timeout: TIMEOUT)
         if !iPad() {
             app.buttons["Close"].tap()
+        } else {
+            // Workaround to close the context menu.
+            // XCUITest does not allow me to click the greyed out portion of the app.
+            app.otherElements["ActivityListView"].cells["XCElementSnapshotPrivilegedValuePlaceholder"].firstMatch.tap()
+            waitForExistence(app.alerts["Unable to Import Clinical Documents"], timeout: TIMEOUT)
+            app.buttons["OK"].tap()
         }
      }
 
@@ -124,18 +150,18 @@ class DownloadFilesTests: BaseTestCase {
         waitUntilPageLoad()
         app.webViews.firstMatch.swipeLeft()
         for _ in 0..<numberOfDownlowds {
-            waitForExistence(app.webViews.links[testFileName], timeout: 5)
+            waitForExistence(app.webViews.links[testFileName], timeout: TIMEOUT)
 
             app.webViews.links[testFileName].firstMatch.tap()
 
-            waitForExistence(app.tables["Context Menu"].otherElements["download"], timeout: 5)
+            waitForExistence(app.tables["Context Menu"].otherElements["download"], timeout: TIMEOUT)
             app.tables["Context Menu"].otherElements["download"].tap()
         }
     }
 
     private func downloadBLOBFile() {
         navigator.openURL(testBLOBURL)
-        waitForExistence(app.webViews.links["Download Text"], timeout: 5)
+        waitForExistence(app.webViews.links["Download Text"], timeout: TIMEOUT)
         app.webViews.links["Download Text"].press(forDuration: 1)
         app.buttons["Download Link"].tap()
     }
@@ -168,7 +194,7 @@ class DownloadFilesTests: BaseTestCase {
 
         // Remove private data once the switch to remove downloaded files is enabled
         navigator.goto(NewTabScreen)
-        waitForExistence(app.buttons["urlBar-cancel"], timeout: 5)
+        waitForExistence(app.buttons["urlBar-cancel"], timeout: TIMEOUT)
         navigator.performAction(Action.CloseURLBarOpen)
         navigator.nowAt(NewTabScreen)
         navigator.goto(ClearPrivateDataSettings)
@@ -176,7 +202,7 @@ class DownloadFilesTests: BaseTestCase {
         navigator.performAction(Action.AcceptClearPrivateData)
 
         navigator.goto(HomePanelsScreen)
-        waitForExistence(app.buttons["urlBar-cancel"], timeout: 5)
+        waitForExistence(app.buttons["urlBar-cancel"], timeout: TIMEOUT)
         navigator.performAction(Action.CloseURLBarOpen)
         navigator.nowAt(NewTabScreen)
         navigator.goto(LibraryPanel_Downloads)
@@ -185,7 +211,7 @@ class DownloadFilesTests: BaseTestCase {
     }
 
     private func checkTheNumberOfDownloadedItems(items: Int) {
-        waitForExistence(app.tables["DownloadsTable"], timeout: 10)
+        waitForExistence(app.tables["DownloadsTable"], timeout: TIMEOUT)
         let list = app.tables["DownloadsTable"].cells.count
         XCTAssertEqual(list, items, "The number of items in the downloads table is not correct")
     }
@@ -194,7 +220,7 @@ class DownloadFilesTests: BaseTestCase {
         downloadFile(fileName: testFileName, numberOfDownlowds: 1)
         waitForExistence(app.buttons["Downloads"])
         app.buttons["Downloads"].tap()
-        waitForExistence(app.tables["DownloadsTable"], timeout: 3)
+        waitForExistence(app.tables["DownloadsTable"], timeout: TIMEOUT)
         checkTheNumberOfDownloadedItems(items: 1)
     }
 }
